Removed original root dir contents from tracking.
Basically, I've modified the repo so that it now tracks the parent dir
of the original repo root dir, mycoderepo, which is now a subdir of the
new repo root dir.
As a result, I have to "remove" what git still thinks are the root
folder contents (now instead within mycoderepo).